% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDim.R
\name{addDim}
\alias{addDim}
\title{adds a dimension to a table - for testing purposes on cross joins}
\usage{
addDim(DT, dim, dimName)
}
\arguments{
\item{DT}{the DataTable you want to manipulate}

\item{dim}{unique values you want your table to be duplicated over}

\item{dimName}{name of the new column}
}
\description{
add dim to a table
}
