% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_pad.R
\name{cols_pad}
\alias{cols_pad}
\title{add variables so that data can be easily used in a pivot}
\usage{
cols_pad(data, nCols = 100, colPrefix = "x_")
}
\arguments{
\item{data}{the data to add columns to}

\item{nCols}{number of columns to define}

\item{colPrefix}{for the dummy columns}
}
\value{
the table with the chosen number of columns
}
\description{
will simply add dummy columns up to the number defined by the
user
}
