% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLibsR.AlphA.R
\name{getLibsR.AlphA}
\alias{getLibsR.AlphA}
\title{Library Loader}
\usage{
getLibsR.AlphA()
}
\value{
The value from the last call 'library' call in the function.
}
\description{
Loads all libraries in a pre defined set, only for code
wrapping purposes.
}
