% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listVarsUniques.R
\name{listVarsUniques}
\alias{listVarsUniques}
\title{List Unique Values of Columns Across Tables}
\usage{
listVarsUniques(..., pattern = "")
}
\arguments{
\item{...}{One or more data.tables, or a list of data.tables, to process.}

\item{pattern}{An optional regular expression pattern to filter column names.}
}
\value{
A data.table summarizing the number of unique values for each column
in each table.
}
\description{
Counts the unique values of columns across one or more
data.tables, optionally filtered by a pattern.
}
\examples{
library(data.table)

table1 <- data.table(a = c(1, 2, 3, 4, 5))
table2 <- data.table(b = c(2, 2, 3, 3, 3))
tablesList <- list(table1=table1, table2=table2)
listVarsUniques(tablesList, pattern = "^a")
}
