% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCheck.R
\name{loadCheck}
\alias{loadCheck}
\title{Load and Install Package if Necessary}
\usage{
loadCheck(package_names)
}
\arguments{
\item{package_names}{A character vector specifying the name(s) of the
package(s) to install (if necessary), and load.}
}
\value{
No return value.
}
\description{
This function checks if a specified package is available in the
current R environment. If the package is not installed, it automatically
installs it with dependencies and then loads it. #' The function suppresses
startup messages to provide a clean loading experience.
}
\examples{
# Load a commonly used package
loadCheck("dplyr")

# Load a package that might not be installed
loadCheck("ggplot2")
}
