% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lum_0_100.R
\name{lum_0_100}
\alias{lum_0_100}
\title{Adjust the Brightness of the Graphics Window}
\usage{
lum_0_100(lum)
}
\arguments{
\item{lum}{Numeric. Brightness level, ranging from 0 (completely dark)
to 100 (maximum brightness).}
}
\description{
Modifies the brightness level of the active graphics window by
adjusting its background color.
}
