% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namedList.R
\name{namedList}
\alias{namedList}
\title{Create a Named List from a Character Vector}
\usage{
namedList(vec)
}
\arguments{
\item{vec}{A character vector.}
}
\value{
A named list where each element is named after its value.
}
\description{
Converts a character vector into a named list where each
element's name is its own value.
}
\examples{
namedList(c("Apple", "Banana", "Cherry"))
}
