% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickSave.R
\name{quickSave}
\alias{quickSave}
\title{Save File in a Directory storing saves, prefixing it with current date}
\usage{
quickSave(
  saveDir,
  filePath = NULL,
  saveNote = NULL,
  overwrite = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{saveDir}{Choose the directory used to store saves. Suggested : 'old'}

\item{filePath}{Optional, if you want to save another file than the current
one : full path of the file you want to save.}

\item{saveNote}{An optional custom note to append to the file name
for the save, allowing to keep track of why this save has been done.}

\item{overwrite}{Logical. Should an existing save with the same name
be overwritten? Default is `FALSE`.}

\item{verbose}{logical. If turned to `TRUE`, the save path is displayed}
}
\value{
the output value of the function used to copy file
}
\description{
Saves a file with current date in its name in a sub directory
located in the same directory as the original file.
Optionally, a note is added after the file name.
}
