% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ret_lum.R
\name{ret_lum}
\alias{ret_lum}
\title{change luminosity of a hex color by multiplying rgb components by a factor}
\usage{
ret_lum(hexCol, rgbFact)
}
\arguments{
\item{hexCol}{color to retreat, in hexaDecimal format}

\item{rgbFact}{factor to multiply rgb colors with. if > 1, check final value}
}
\description{
change luminosity of a hex color by multiplying rgb components by a factor
}
