% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBBGColor_lum_0_100.R
\name{SBBGColor_lum_0_100}
\alias{SBBGColor_lum_0_100}
\title{Set Shiny Background Color to a Chosen Shade of Grey}
\usage{
SBBGColor_lum_0_100(lum)
}
\arguments{
\item{lum}{Numeric. Luminosity level, ranging from 0 (black) to 100 (white).}
}
\value{
An HTML tag representing the background color for the sidebar.
}
\description{
Adjusts the background color of a Shiny app's sidebar based
on a specified luminosity.
}
