% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sepThsd.R
\name{sepThsd}
\alias{sepThsd}
\title{Quick formatting}
\usage{
sepThsd(x, big.mark = " ", digits = 1, scientific = FALSE)
}
\arguments{
\item{big.mark}{separator for thousands}

\item{digits}{number of digits}

\item{scientific}{boolean}
}
\description{
wrapper of format function with custom defaults
}
