% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_lum_0_100.R
\name{shiny_lum_0_100}
\alias{shiny_lum_0_100}
\title{set the Shiny BG color to chosen shade of grey}
\usage{
shiny_lum_0_100(lum)
}
\arguments{
\item{lum}{luminosity between 0 and 100}
}
\value{
the different HTML tags for background, sidebar, (+ more ?)
}
\description{
choose luminosity between 0 and 100
}
