% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slicetr.R
\name{slicetr}
\alias{slicetr}
\title{Quick Look at the First Rows of a Table}
\usage{
slicetr(x, n = 3)
}
\arguments{
\item{x}{Data frame or table. The input table to inspect.}

\item{n}{Integer. Number of rows to display. Default is 3.}
}
\value{
None. The function prints the formatted rows to the console.
}
\description{
Displays the first rows of a table with vertically aligned
values, offering a clearer view than `str()`.
}
