% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_table.R
\name{time_table}
\alias{time_table}
\title{Times steps}
\usage{
time_table(tt_message, tt_reset = FALSE, ...)
}
\arguments{
\item{tt_message}{mandatory}

\item{tt_reset}{to start a new time table from scratch}

\item{...}{any other specification. Choose name for column and value for row}
}
\value{
nothing, creates a dt in parent environment : maybe not the best practice
}
\description{
previous version - deprecated - use timer
}
\examples{
time_table("step 1")
time_table("step 2")
time_table("starting new cycle", lines_volume = 100)
time_table("starting new cycle", lines_volume = 1000)
}
