% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmr_compute.R
\name{tmr_compute}
\alias{tmr_compute}
\title{Times steps - v3 with lubridate functions, and only stores time, not durations}
\usage{
tmr_compute(tmTbl, starts = TRUE)
}
\arguments{
\item{tmTbl}{an existing time table}

\item{ends}{do timers correspond to ends of steps}
}
\value{
the tmr table sorted and with computed columns
}
\description{
computing columns in a tmr table
}
\examples{
new_tmr <- tmr_init(step = 0)
new_tmr <- tmr(new_tmr, step = 1)
full_tmr <- tmr_compute(new_tmr)
}
