% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimmid.R
\name{trimmid}
\alias{trimmid}
\title{Tronquer une chaîne au milieu}
\usage{
trimmid(x, n = 20)
}
\arguments{
\item{x}{Vecteur character.}

\item{n}{Longueur max (défaut 20).}
}
\value{
Character tronqué si nécessaire.
}
\description{
Garde le début et la fin, remplace le milieu par "...".
Vectorisé.
}
