% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_pad.R
\name{cols_pad}
\alias{cols_pad}
\title{Add Variables to Facilitate Data Usage in a Pivot Table}
\usage{
cols_pad(data, nCols = 100, colPrefix = "x_")
}
\arguments{
\item{data}{The data frame to which dummy columns will be added.}

\item{nCols}{An integer specifying the total number of columns desired.}

\item{colPrefix}{A string used as the prefix for the names of dummy columns.}
}
\value{
A data frame with the specified total number of columns.
}
\description{
Adds dummy columns to reach the number specified by the user.
}
\examples{
table <- data.frame(a = 1:5, b = letters[1:5])
extraTable <- cols_pad(table, nCols = 6, colPrefix = "extra_")
print(extraTable)
}
