% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareVars.R
\name{compareVars}
\alias{compareVars}
\title{Compare tables variables}
\usage{
compareVars(x, y, pattern = "")
}
\arguments{
\item{x}{first table}

\item{y}{second table}

\item{pattern}{to compare only a subset of all the varibles}
}
\value{
an object containing :
all : all the column names
common : names found in both tables
onlyX : only in the first table
onlyY : obvious
exclusive : columns found in only one of the two tables
}
\description{
reads names matching a given pattern in two tables, and gives
information about which columns are in both tables, only in x/y and so on
}
