% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countSwitches.R
\name{countSwitches}
\alias{countSwitches}
\title{Count Series Based on Start/Stop Markers}
\usage{
countSwitches(
  data,
  colNm,
  sttMark,
  endMark,
  includeStt = TRUE,
  includeEnd = TRUE
)
}
\arguments{
\item{data}{A data frame containing the column to process.}

\item{colNm}{A string specifying the column name in `data` to evaluate.}

\item{sttMark}{A value indicating the start of a series.}

\item{endMark}{A value indicating the end of a series.}

\item{includeStt}{Logical. Should the start marker be included as part of
the series? Default is `TRUE`.}

\item{includeEnd}{Logical. Should the end marker be included as part of the
series? Default is `TRUE`.}
}
\value{
A modified version of the input data frame with additional
columns for series identification and counts.
}
\description{
This function aims at identifying sections and sub-sections
numbers, based on markers of section starts and ends.

Given a data frame, and a column name, it will add columns giving infos
about the categories

.... to be completed ...
 identifies and counts series in a vector based on
specified start/end markers. It returns a vector of the same length,
indicating the series count or 0 when the element is outside a series
(e.g., after a stop marker and before the next start marker).
}
