% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countSwitches.R
\name{countSwitches}
\alias{countSwitches}
\title{Count Series Based on Start/Stop Markers}
\usage{
countSwitches(
  data,
  colNm,
  sttMark,
  endMark,
  includeStt = TRUE,
  includeEnd = TRUE
)
}
\arguments{
\item{data}{A data frame containing the column to process.}

\item{colNm}{A string specifying the column name in `data` to evaluate.}

\item{sttMark}{A value indicating the start of a series.}

\item{endMark}{A value indicating the end of a series.}

\item{includeStt}{Logical. Should the start marker be included as part of
the series? Default is `TRUE`.}

\item{includeEnd}{Logical. Should the end marker be included as part of the
series? Default is `TRUE`.}
}
\value{
A modified version of the input data frame with additional columns including:
\itemize{
  \item `stepStr`: The column specified by `colNm`.
  \item `findStt`: Logical flag indicating the occurrence of the start marker.
  \item `findEnd`: Logical flag indicating the occurrence of the end marker.
  \item `catLvl`: The current series level calculated as the difference between the cumulative counts of start and end markers.
  \item `inc1`, `inc2`, `inc3`: Logical flags for detecting nested series at different levels.
  \item `check1`, `check2`, `check3`: Intermediate variables used for series counting.
  \item `tst1`, `tst2`, `tst3`: Cumulative series counts at different nesting levels.
  \item `ret1`, `ret2`, `ret3`: Final series counts returned for each respective level.
}
}
\description{
This function identifies and counts series in a vector based on
specified start and stop markers. It returns a vector of the same length,
indicating the series count or 0 when the element is outside a series
(e.g., after a stop marker and before the next start marker).
}
