% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAB_checkPbs.R
\name{FAB_checkPbs}
\alias{FAB_checkPbs}
\title{FAB_checkPbs}
\usage{
FAB_checkPbs(linesBefore = 0, linesAfter = 10)
}
\arguments{
\item{linesBefore}{nb of lines before the pb to display}

\item{linesAfter}{nb of lines after the pb to display}
}
\description{
gives a glimpse of lines which could cause pbs in FAB
}
