% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldAllBr.R
\name{foldAllBr}
\alias{foldAllBr}
\title{Easily Fold Code Parts}
\usage{
foldAllBr(time = FALSE, debug_getTbl = FALSE)
}
\arguments{
\item{time}{Logical. If `TRUE`, the function will provide timing information
for each step.}

\item{debug_getTbl}{Logical. If `TRUE`, returns the `docContent` table with
tags for debugging purposes.}
}
\value{
If `debug_getTbl = TRUE`, a table (`docContent`) with tags.
Otherwise, no return value (performs folding).
}
\description{
This function works with code split into parts identified by brackets.
The format is as follows:
\preformatted{
\{
  ...
  code from part 1
  ...
\} # part 1
\{
  ...
\} # part 2
}

It automatically identifies parts to fold/unfold easily.

Shortcuts required:
\itemize{
  \item "fold all brackets": shift + alt + S (Windows) / ctrl + shift + up (Mac)
  \item "expand fold": shift + alt + D (Windows) / ctrl + shift + down (Mac)
}
}
