% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldAllBr.R
\name{foldAllBr}
\alias{foldAllBr}
\title{Fold All Brackets in Code}
\usage{
foldAllBr(time = F, debug_getTbl = F)
}
\arguments{
\item{time}{Logical. If `TRUE`, the function will provide timing information
for each step.}

\item{debug_getTbl}{Logical. If `TRUE`, returns the `docContent` table with
tags for debugging purposes.}
}
\value{
If `debug_getTbl = TRUE`, a table (`docContent`) with tags.
Otherwise, no return value (performs folding).
}
\description{
Automatically identifies open brackets in the current code and
folds them.
}
