% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeCols.R
\name{initializeCols}
\alias{initializeCols}
\title{initialize columns}
\usage{
initializeCols(data, cols, value)
}
\arguments{
\item{data}{table where you need to init cols}

\item{cols}{col names}

\item{value}{the value you want to initialize if column doesn't exist yet}
}
\value{
directly change data by reference (DT way, maybe to be changed)
}
\description{
check if columns are present - if not create them with dft value
}
