% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeCols.R
\name{initializeCols}
\alias{initializeCols}
\title{Initialize Missing Columns in Data Table}
\usage{
initializeCols(data, cols, value)
}
\arguments{
\item{data}{A data.table to modify. Converts to data.table if necessary.}

\item{cols}{Character vector of column names to check and initialize.}

\item{value}{Default value to assign to missing columns.}
}
\value{
The modified data.table, updated by reference.
}
\description{
Adds missing columns to a data.table and initializes them
with a default value.
}
