% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_join_checks.R
\name{left_join_checks}
\alias{left_join_checks}
\title{Left Join with Validation Checks}
\usage{
left_join_checks(
  x,
  y,
  ...,
  req_xAllMatch = 1,
  req_preserved_x = 1,
  behavior = "error",
  showNotFound = F,
  time = F
)
}
\arguments{
\item{x}{A data.table representing the left table.}

\item{y}{A data.table representing the right table.}

\item{...}{Additional arguments passed to `dplyr::left_join`.}

\item{req_xAllMatch}{Logical. Ensure that all rows in `x` find a match in `y`. Default: FALSE.}

\item{req_preserved_x}{Logical. Ensure that the number of rows in `x`
remains unchanged after the join. Default: TRUE.}

\item{behavior}{Character. Specifies behavior if validation fails.
Options: `"warning"` or `"error"`. (default: `"warning"`)}

\item{showNotFound}{Logical. Show rows from `x` that did not match with `y`.
Default: FALSE.}

\item{time}{Logical. Internal argument used only for testing purposes during
manual runs.}
}
\value{
A data.table containing the joined table.
}
\description{
Performs a left join and verifies that no unexpected duplicates
or mismatches occur.
}
