% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_join_checks.R
\name{left_join_checks}
\alias{left_join_checks}
\title{left_join + checks if everything went correctly}
\usage{
left_join_checks(
  x,
  y,
  ...,
  req_xAllMatch = 1,
  req_preserved_x = 1,
  behavior = "error",
  showNotFound = F
)
}
\arguments{
\item{x}{left table}

\item{y}{right table}

\item{...}{additionnal arguments passed to dplyr's left_join}

\item{req_xAllMatch}{boolean : do you want to ensure that all x values have found a match in y ?}

\item{req_preserved_x}{boolean : do you want to ensure that xTable rows are the same after join (no duplicates) ?}

\item{behavior}{character : warning, or error}

\item{showNotFound}{if some x are not found in y, do you want to show them}
}
\value{
the joined table
}
\description{
typically, check that you did not duplicate your lines
}
