% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listVarsUniques.R
\name{listVarsUniques}
\alias{listVarsUniques}
\title{listing dimensions}
\usage{
listVarsUniques(..., pattern = "")
}
\arguments{
\item{...}{should be a lsit a tables}
}
\description{
list unique values of columns for a set of tables
}
