% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lum_0_100.R
\name{lum_0_100}
\alias{lum_0_100}
\title{regler la luminosité de la fenetre graphique}
\usage{
lum_0_100(lum)
}
\arguments{
\item{lum}{luminosity from 0 to 100}
}
\description{
regler la luminosité de la fenetre graphique
}
