% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printif.R
\name{printif}
\alias{printif}
\title{Conditionally Print an Object}
\usage{
printif(x, show = FALSE, ...)
}
\arguments{
\item{x}{Any object.}

\item{show}{A logical value indicating whether to print `x` (default: `FALSE`).}
}
\value{
The object `x`, possibly printed if `show` is `TRUE`.
}
\description{
This function prints `x` if `show` is `TRUE`;
otherwise, it returns `x` unchanged.
}
\examples{
library(dplyr)
mtcars \%>\%
  filter(mpg > 20) \%>\%
  summarise(mean_hp = mean(hp)) \%>\%
  printif(1)
}
