% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdate.R
\name{rdate}
\alias{rdate}
\title{Random dates}
\usage{
rdate(
  x,
  min = paste0(format(Sys.Date(), "\%Y"), "-01-01"),
  max = paste0(format(Sys.Date(), "\%Y"), "-12-31"),
  sort = FALSE,
  include_hours = FALSE
)
}
\arguments{
\item{x}{length of the vector}

\item{min}{opt : minimum date}

\item{max}{opt : maximum date}

\item{sort}{do you want sorted dates, default FALSE}

\item{include_hours}{should returned dates include time also (default FALSE : days only)}
}
\value{
a dates vector of length x
}
\description{
get a vector of x random dates between min and max
}
\examples{
rdate(10)
}
