% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readable_period.R
\name{readable_period}
\alias{readable_period}
\title{Convert Period Object to Human-Readable Text}
\usage{
readable_period(periodObj)
}
\arguments{
\item{periodObj}{A `period` object from the `lubridate` package to format.}
}
\value{
A character string representing the period in a human-readable format.
}
\description{
Converts a `lubridate` period object into a human-readable
string representation.
}
\examples{
library(lubridate)

periodObj <- days(2) + hours(3) + minutes(15) + seconds(30)
print(periodObj)
readable_period(periodObj)
}
