% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root.R
\name{root}
\alias{root}
\title{Get Root Directory of Current Source File}
\usage{
root()
}
\value{
A character string representing the absolute path of the directory
containing the current source file.
}
\description{
Returns the directory path where the current source code file
is located.
}
