% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_in_old.R
\name{save_in_old}
\alias{save_in_old}
\title{Save File in an "Old" Directory, prefixing it with current date}
\usage{
save_in_old(
  saveDir,
  filePath = NULL,
  saveNote = NULL,
  overwrite = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{saveDir}{Choose the directory used to store saves. for instance : 'old'}

\item{filePath}{Optional, if you want to save another file than the current
one : full path of the file you want to save.}

\item{saveNote}{An optional custom note to append to the file name
for the save, allowing to keep track of why this save has been done.}

\item{overwrite}{Logical. Should an existing save with the same name
be overwritten? Default is `FALSE`.}

\item{verbose}{logical. If turned to `TRUE`, the save path is displayed}
}
\value{
the output value of the function used to copy file
}
\description{
Saves a file with the current date in its name in an "old"
directory located in the same directory as the original file.
Optionally, a note is added after the file name.
}
