% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBBGColor_lum_0_100.R
\name{SBBGColor_lum_0_100}
\alias{SBBGColor_lum_0_100}
\title{set the Shiny BG color to chosen shade of grey}
\usage{
SBBGColor_lum_0_100(lum)
}
\arguments{
\item{lum}{luminosity between 0 and 100}
}
\value{
the HTML tag for sidebar color
}
\description{
choose luminosity between 0 and 100
}
