% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setOption.R
\name{setOption}
\alias{setOption}
\title{Sets a global option from a named list element}
\usage{
setOption(optionList)
}
\arguments{
\item{optionList}{An element from a named list, specified as `myList$element`.}
}
\value{
The function does not return anything but sets an option that can be retrieved
using `getOption(list_name)`.
}
\description{
This function takes an element from a named list as an argument
and sets a global option where the option name is the list's name and the
value is the corresponding string of the selected element.
}
\details{
- The function automatically extracts the list name from the argument.
- The option is then dynamically set using `options(list_name = element)`.
}
\examples{
myList <- list(model1 = "model_1", model2 = "model_2", model3 = "model_3")
setOption(myList$model1)
getOption("myList")  # Returns "model_1"

}
