% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_lum_0_100.R
\name{shiny_lum_0_100}
\alias{shiny_lum_0_100}
\title{Set Shiny Background and Sidebar Colors to a Chosen Shade of Grey}
\usage{
shiny_lum_0_100(lum)
}
\arguments{
\item{lum}{Numeric. Luminosity level, ranging from 0 (black) to 100 (white).}
}
\value{
The HTML tags for setting the background and sidebar colors.
}
\description{
Adjust the background color of a Shiny app's main body and sidebar
based on a specified luminosity level.

The purpose is the same as lum_0_100() function, avoiding problems with high
contrast between with graphic windows and dark themes.
}
