% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slicetr.R
\name{slicetr}
\alias{slicetr}
\title{quick look at first rows}
\usage{
slicetr(x, n = 3)
}
\arguments{
\item{x}{table}

\item{n}{number of rows to display}
}
\value{
nothing
}
\description{
same doal as str but with vertically aligned values
}
