% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortBy.R
\name{sortby}
\alias{sortby}
\title{Sort a Data Frame}
\source{
https://towardsdatascience.com/creating-a-custom-r-package-9a2e303d3332
}
\usage{
sortby(df, col, desc = FALSE)
}
\arguments{
\item{df}{The df to be sorted}

\item{col}{The column name (string) by which to sort the df}

\item{desc}{self explained, default FALSE}
}
\value{
The sorted df
}
\description{
This function accepts a df and col number to sort the df by that column
}
\examples{
subject <- c(1:10)
height <- round(rnorm(10, mean = 66, sd = 3), 1)
d <- data.frame(subject, height)

d2 <- sortby(d, "height")
d2

d3 <- sortby(d, "height", desc = TRUE)
d3


}
