% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\name{timer}
\alias{timer}
\title{Times steps - new version : don't create time table as global variable cause it sucks}
\usage{
timer(timer_table = data.table(), start = FALSE, message = F, ...)
}
\arguments{
\item{timer_table}{table to increment with a new timer line}

\item{start}{to start a new time table from scratch}

\item{message}{Should the function print the table for this step ?}

\item{...}{any other specification. Choose name for column and value for row}
}
\value{
the provided timer_table plus one line with time and the ... specifications
}
\description{
keep track of time spent on different steps of code
}
\examples{
tt_tests <- timer(start = TRUE)
tt_tests <- timer(timer_table = tt_tests)
tt_tests <- timer(timer_table = tt_tests, stepName = "step x", anyVariableName = "any value")
}
