% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmr.R
\name{tmr}
\alias{tmr}
\title{Times steps - v3 with lubridate functions, and only stores time, not durations}
\usage{
tmr(tmTbl, message = F, ...)
}
\arguments{
\item{tmTbl}{an existing time table}

\item{message}{print the timer}

\item{...}{any other specification. Choose name for column and value for row}
}
\value{
the tmr table with 1 line added at the end
}
\description{
initialisation of a tmr_table
}
\examples{
new_tmr <- tmr_init(step = 0)
new_tmr <- tmr(new_tmr, step = 1)
}
